Add watched parameter to gatherMovieData
Updated gatherMovieData function to accept a boolean parameter 'watched' indicating if the movie has been watched.
This change allows for more flexibility in handling movie data based on its watched status.
Added support for the 'watched' parameter in the gatherMovieData function and POST handler.
Introduced a new Checkbox component for UI interaction with the 'watched' field in the MovieWidget form.